Only the carbonate output grids are in the new reference frame.
This fixes 0f9ac4e which presumed the age and bathymetry grids were in
the new reference frame - they should instead continue to be sampled in
the default reference frame. Essentially the carbonate grids are rotated
into the new reference frame but only keeping the regional lat/lon
extent (if specified).